% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap-print.R
\docType{methods}
\name{print.Bootstrap}
\alias{print,Bootstrap-method}
\alias{print.Bootstrap}
\title{\code{\link{print}} or \code{\link{summary}} a Bootstrap object}
\usage{
\S4method{print}{Bootstrap}(x)
}
\arguments{
\item{x}{Bootstrap class object.}
}
\value{
according to the call
 \item{print}{console output text with increasing detail of
 Bootstrap object.}
}
\description{
Generic Print/Summary method for Bootstrap class output visualization.
}
\examples{
{
 data(NCI60Selec)
 fit<-DiStatis(NCI60Selec,Scale=TRUE,Center=TRUE)
 A<-Bootstrap(fit)

 print(A)


}
}
\author{
M L Zingaretti,  J A Demey-Zambrano, J L Vicente-Villardon, J R Demey
}

