% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap-print.R
\docType{methods}
\name{summary.Bootstrap}
\alias{summary,Bootstrap-method}
\alias{summary.Bootstrap}
\title{\code{\link{print.Bootstrap}} or \code{\link{summary.Bootstrap}} a Bootstrap object}
\usage{
\S4method{summary}{Bootstrap}(object)
}
\arguments{
\item{object}{Bootstrap class object.}
}
\value{
according to the call
 \item{ summary}{console output text with increasing detail of
 Bootstrap object.}
}
\description{
Generic Print/Summary method for Bootstrap class output visualization.
}
\examples{
{
 data(NCI60Selec)
 fit<-DiStatis(NCI60Selec,Scale=TRUE,Center=TRUE)
 A<-Bootstrap(fit)


 summary(A)

}
}
\author{
M L Zingaretti, J A Demey, J L Vicente-Villardon, J R Demey
}

