% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiStatis-Bootstrap.R
\docType{methods}
\name{Bootstrap}
\alias{Bootstrap}
\alias{Bootstrap,DiStatis-method}
\alias{Bootstrap-methods}
\title{Funtion \code{Bootstrap} of a DiStatis object}
\usage{
Bootstrap(object, NRep = 100, Dims = 2)

\S4method{Bootstrap}{DiStatis}(object, NRep = 100, Dims = 2)
}
\arguments{
\item{object}{It is the object of DiStatis Class.}

\item{NRep}{Number of repetitions to do the bootstraping. Default is 100.}

\item{Dims}{Number of dimensions used to do the bootstraping. Default is 2.}
}
\value{
\item{Bootstrap}{Bootstrap class object with the
  corresponding completed slots
  according to the given model}
}
\description{
This function is use to make Bootstrap from DiStatis object. Bootstrap
resampling techniques are applied on the residuals
matrices obtained from SVD of the Compromise and
the RV matrix to do multiple comparisions between
studies and  confidence elipses for the projections
of observations in the compromise.
}
\examples{
{
data(NCI60Selec)
Z1<-DiStatis(NCI60Selec)
B<-Bootstrap(Z1)
Z2<-DiStatis(NCI60Selec,Center=FALSE,Scale=FALSE)
B2<-Bootstrap(Z2,NRep=200)

}
}
\author{
M L Zingaretti, J A Demey-Zambrano, J L Vicente Villardon, J R Demey
}
\references{
\enumerate{
 \item Efron, B.,Tibshirani, RJ. (1993). An introduction to the bootstrap. New York: Chapman and Hall. 436p.
 \item Ringrose, T.J. (1992). Bootstrapping and Correspondence Analysis in Archaeology. Journal of Archaeological. Science.19:615-629.

}
}

