% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiStatis-CompPlot.R
\docType{methods}
\name{CompPlot}
\alias{CompPlot}
\alias{CompPlot,DiStatis-method}
\title{Plot a \code{CompPlot} of a DiStatis object}
\usage{
\S4method{CompPlot}{DiStatis}(x,xlabProj=NULL, ylabProj=NULL,
xlabBar="Inertia(\%)",ylabBar="Values",mainBar=NULL,mainProj=NULL,
pchPoints=15,legend=TRUE ,colBar="red",colObs="black",barPlot=TRUE,...)
}
\arguments{
\item{x}{DiStatis class object.}

\item{xlabProj}{character for the x-label title for plot}

\item{ylabProj}{character for the x-label title for plot}

\item{xlabBar}{character for the x-label title for barPlot}

\item{ylabBar}{character for the y-label title for barPlot}

\item{mainBar}{main to histogram plot}

\item{mainProj}{main to proj plot}

\item{pchPoints}{pch for points in plot.}

\item{legend}{Logical. indicates whether the legend prints}

\item{colBar}{character col for bars in the BarPlot}

\item{colObs}{character col for observations in the plot}

\item{barPlot}{logical indicates whether the barPlot is prints}

\item{...}{additional parameters for plot}
}
\value{
plotted CompPlot/s of the component/s of the given DiStatis object.
}
\description{
Plot a \code{CompPlot} of a DiStatis object
}
\examples{
{
\dontrun{
 data(NCI60Selec)
 Z2<-DiStatis(NCI60Selec,Scale=TRUE,Center=TRUE)
 Colores2<-c(rep(colors()[657],5),rep(colors()[637],6),
 rep(colors()[537],7),rep(colors()[552],6),rep(colors()[57],10),
 rep(colors()[300],9),rep(colors()[461],7),rep(colors()[450],2),
 rep(colors()[432],8))
 Colores1<-c(rep("Breast",5),rep("CNS",6),rep("Colon",7),
 rep("Leukemia",6),rep("Melanoma",10),rep("Lung",9),rep("Ovarian",7),
 rep("Prostate",2),rep("Renal",8))
 CompPlot(Z2,xlabBar="",colObs=Colores2,pch=15,las=1,
 cex=2,legend=FALSE,barPlot=FALSE,cex.main=0.6,cex.lab=0.6,
 cex.axis=0.6,las=1)
 legend("topleft",unique(Colores1),col=unique(Colores2),
 bty="n",pch=16,cex=1)
}
}
}
\author{
M L Zingaretti,  J A Demey-Zambrano, J L Vicente Villardon, J R Demey
}

