% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiStatis-DiStatis.R
\docType{methods}
\name{DiStatis}
\alias{DiStatis}
\alias{DiStatis-methods}
\title{\code{DiStatis} of a DiStatis object
High level constructor of DiStatis class object}
\format{An object of class \code{NULL} of length 0.}
\usage{
DiStatis(Data = NULL, Distance = c(), Center = TRUE, Scale = TRUE,
  CorrelVector = TRUE, Frec = FALSE, Traj = TRUE)
}
\arguments{
\item{Data}{The data frame or of k-tables type. The Observations should be in rows (common elements in DAnisostatis), the variables and Studies must be in columns. After the name of the variable an underscore (_) must be written to indicate the Study (eg. Var1_Est1 , eg. Var1_EstK, for more information see the data object). The name of a variable can include any symbol except an underscore (_). REMEMBER the underscore (_) should be reserved to indicate the study.
Also, the Data can be a list of k components.
 Each element of the list is one of the tables
 with observations in rows and variables in columns.
 The elements of list must be data.frame or ExpressionSet data.}

\item{Distance}{Vector is the length equal to the number of studies
that indicates the kind of distance (or scalar product)
that is calculated in each study. If not specify
(or is wrong  specify) the scalar product is used.
The options can be ScalarProduct, euclidean,
manhattan, canberra, pearson, pearsonabs, spearman,
spearmanabs, mahalanobis. In the binary data the distance can be: jaccard,
simple matching, sokal&Sneath, Roger&Tanimoto, Dice,
Hamman,#'  Ochiai, Sokal&Sneath, Phi-Pearson,
Gower&Legendre.}

\item{Center}{A logical value. If TRUE, the data frame
is centered  by the mean. By default is TRUE.}

\item{Scale}{A logical value indicating whether the column vectors (of the data.frame) should be
standardized by the rows weight, by default is TRUE.}

\item{CorrelVector}{a logical value. If TRUE (default), Vectorial correlation coefficient  is computed for the  RV matrix.
If FALSE the Hilbert-Smith distance is used in the RV matrix.}

\item{Frec}{Logical. Should the data be treated data as frequencies? By default is FALSE.}

\item{Traj}{Logical. Should the trajectories analysis be done?  By default is TRUE.}
}
\value{
\item{DiStatis}{DiStatis class object with the
  corresponding completed slots
 according to the given model}
}
\description{
This is the function that makes DiStatis Methodology: Statis is part of the PCA
family and therefore the main analytical tool for STATIS
is the singular value decomposition (SVD) and the
generalized singular value decomposition (GSVD) of a matrix.
The goal of Statis is to analyze several data sets of
variables that were collected on the same set of observations.
Originally, the comparisons were drawn from the compute
of the scalar product between the different tables.
In this approach, the condition is made more flexible,
allowing  the incorporation of different distance measurements
(including the scalar product)  to compare the tables.

function to do Statis (K-tables methodology)with distance options
}
\details{
STATIS methods: to more information, see references.
}
\section{Note}{
 use \bold{\code{\link{DiStatis-class}}} high level
constructor for the
 creation of the class instead of directly
 calling its constructor by new
 means.
}
\examples{
{
data(NCI60Selec_ESet)
Z1<-DiStatis(NCI60Selec_ESet)
data(winesassesors)
Z3<-DiStatis(winesassesors)

}

}
\author{
M L Zingaretti,  J A Demey-Zambrano, J L Vicente Villardon, J R Demey
}
\references{
\enumerate{
 \item Abdi, H., Williams, L.J.,
 Valentin, D., & Bennani-Dosse, M. (2012).
 STATIS and DISTATIS: optimum multitable principal
 component analysis and three way metric multidimensional scaling.
 WIREs Comput Stat, 4, 124-167.
 \item Escoufier, Y. (1976). Operateur associe a un tableau de donnees. Annales de laInsee, 22-23, 165-178.
 \item Escoufier, Y. (1987). The duality diagram: a means for better practical applications. En P. Legendre & L. Legendre (Eds.), Developments in Numerical Ecology, pp. 139-156, NATO Advanced Institute, Serie G. Berlin: Springer.
 \item L'Hermier des Plantes, H. (1976). Structuration des Tableaux a Trois Indices de la Statistique. [These de Troisieme Cycle]. University of Montpellier, France.


}
}
\keyword{datasets}

