% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiStatis-PanelPlot.R
\docType{methods}
\name{PanelPlot}
\alias{PanelPlot}
\alias{PanelPlot,DiStatis-method}
\title{Plot a \code{PanelPlot} of a DiStatis object}
\usage{
\S4method{PanelPlot}{DiStatis}(x)
}
\arguments{
\item{x}{DiStatis class object.}
}
\value{
Panel plot  of the given DiStatis object.
}
\description{
Plot a \code{PanelPlot} of a DiStatis object
}
\examples{
{
data(NCI60Selec)
Z2<-DiStatis(NCI60Selec,Scale=TRUE,Center=TRUE)
PanelPlot(Z2)
data(winesassesors)
Z3<-DiStatis(winesassesors)
PanelPlot(Z3)

}
}
\author{
M L Zingaretti,  J A Demey-Zambrano, J L Vicente Villardon, J R Demey
}

