% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiStatis-print.R
\docType{methods}
\name{print.DiStatis}
\alias{print,DiStatis-method}
\alias{print.DiStatis}
\title{\code{\link{print.DiStatis}}  a DiStatis object
Generic Print/Summary method for DiStatis class output visualization.}
\usage{
\S4method{print}{DiStatis}(x)
}
\arguments{
\item{x}{DiStatis class object.}
}
\value{
according to the call
 \item{print}{console output text with increasing detail of
 DiStatis object.}
}
\description{
\code{\link{print.DiStatis}}  a DiStatis object
Generic Print/Summary method for DiStatis class output visualization.
}
\examples{
{
 data(NCI60Selec)
 fit<-DiStatis(NCI60Selec,Scale=TRUE,Center=TRUE)

 print(fit)

}
}
\author{
M L Zingaretti,  J A Demey-Zambrano, J L Vicente-Villardon, J R Demey
}
\seealso{
\code{\link{DiStatis}}
}

