% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiStatis-RVPlot.R
\docType{methods}
\name{RVPlot}
\alias{RVPlot}
\alias{RVPlot,DiStatis-method}
\title{Plot a \code{RVPlot} of a DiStatis object}
\usage{
\S4method{RVPlot}{DiStatis}(x,xlabProj="Dim 1", ylabProj="Dim 2",
xlabBar="Inertia(\%)",ylabBar="Values" ,mainBar=NULL,mainProj=NULL,
colArrows="black",legend=TRUE ,colBar="red",barPlot=TRUE,...)
}
\arguments{
\item{x}{DiStatis class object.}

\item{xlabProj}{character for the x-label title for plot}

\item{ylabProj}{character for the y-label title for plot}

\item{xlabBar}{character for the x-label title for barPlot}

\item{ylabBar}{character for the y-label title for barPlot}

\item{mainBar}{the main histogram plot}

\item{mainProj}{the main proj plot}

\item{colArrows}{character col for arrows in ProjPlot}

\item{legend}{Logical. indicates whether the legend prints}

\item{colBar}{character col for bars in the BarPlot}

\item{barPlot}{logical indicates whether the barPlot is prints}

\item{...}{additional parameters for plot}
}
\value{
plotted RVplot/s of the component/s of the given DiStatis object.
}
\description{
Plot a \code{RVPlot} of a DiStatis object
}
\examples{
{
 Z2<-DiStatis(NCI60Selec,Scale=TRUE,Center=TRUE)
 RVPlot(Z2)
 }


}
\author{
M L Zingaretti,  J A Demey-Zambrano, J L Vicente Villardon, J R Demey
}

