% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiStatis-SelectVar.R
\docType{methods}
\name{SelectVar}
\alias{SelectVar}
\alias{SelectVar,DiStatis-method}
\alias{SelectVar-methods}
\title{Function \code{SelectVar} of DiStatis object}
\usage{
SelectVar(object, ord = FALSE, Crit = c("R2-Adj", "p-val(Bonf)", "AIC",
  "BIC"), perc = 0.9, Dims = 2)

\S4method{SelectVar}{DiStatis}(object, ord = FALSE, Crit = c("R2-Adj",
  "p-val(Bonf)", "AIC", "BIC"), perc = 0.9, Dims = 2)
}
\arguments{
\item{object}{Object is an object of DiStatis Class.}

\item{ord}{Logical. If TRUE, the models with intercept are computed, else the intercept is zero.}

\item{Crit}{c("R2-Adj","p-val(Bonf)","AIC","BIC").Criterious of selection. "R2-Adj","p-val (Bonf)","AIC","BIC". Choose "R2-Adj" or "p-val (Bonf)" (Bonferroni correction),"AIC" or "BIC".}

\item{perc}{The value of percentil that indicate how much data than are selected.}

\item{Dims}{Numeric that indicates the number of dimensions to use for do the model. Default is 2.}
}
\value{
\item{SelectVar}{SelectVar class object with the
  corresponding completed slots
  according to the given model}
}
\description{
This function calculates the biplot method through the compromise matrix to select genes
SelectVar from DiStatis Class Object
High level constructor of SelectVar class object
}
\details{
This function allows to build the  biplot for continuous response, using an external procedure
to obtained the regresors in the linear model (the response being an continuous variable).
This function allows the selection of genes using the goodness of fit of the Models Biplot.
object,ord=FALSE,
}
\examples{
{
data(NCI60Selec)
Z1<-DiStatis(NCI60Selec)
M1<-SelectVar(Z1,Crit="R2-Adj",perc=0.95)
M2<-SelectVar(Z1,Crit="p-val(Bonf)",perc=0.95)
}

}
\author{
M L Zingaretti,  J A Demey-Zambrano, J L Vicente Villardon, J R Demey
}
\references{
\enumerate{
 \item Demey, J., Vicente-Villardon, J. L., Galindo, M.P.  & Zambrano, A. (2008) Identifying Molecular Markers Associated With Classification Of Genotypes Using External Logistic Biplots. Bioinformatics, 24(24), 2832-2838.
\item Gabriel, K. (1971). The biplot graphic display of matrices with application to principal component analysis. Biometrika 58(3), 453--467.
\item Gower, J. & Hand, D. (1996). Biplots, Monographs on statistics and applied probability. 54. London: Chapman and Hall., 277 pp.
}
}

