% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiStatis-print.R
\docType{methods}
\name{summary.DiStatis}
\alias{summary,DiStatis-method}
\alias{summary.DiStatis}
\title{\code{\link{summary.DiStatis}}  a DiStatis object}
\usage{
\S4method{summary}{DiStatis}(object)
}
\arguments{
\item{object}{DiStatis class object.}
}
\value{
\item{summary}{console output text with increasing detail of
 DiStatis object.}
}
\description{
\code{\link{summary.DiStatis}}  a DiStatis object
}
\examples{
{
 data(NCI60Selec)
 fit<-DiStatis(NCI60Selec,Scale=TRUE,Center=TRUE)


 summary(fit)

}

}
\author{
M L Zingaretti, J A Demey-Zambrano, J L Vicente-Villardon, J R Demey
}
\seealso{
\code{\link{DiStatis}}
}

