% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiStatis-TrajPlot.R
\docType{methods}
\name{TrajPlot}
\alias{TrajPlot}
\alias{TrajPlot,DiStatis-method}
\title{Plot a \code{TrajPlot} of a DiStatis object}
\usage{
\S4method{TrajPlot}{DiStatis}(x,xlabT="Dim 1", ylabT="Dim 2",
mainTraj=NULL,legend=TRUE ,xlimi=NULL,ylimi=NULL,panel=TRUE,
colours=NULL,...)
}
\arguments{
\item{x}{DiStatis class object.}

\item{xlabT}{character for the x-label title for plot}

\item{ylabT}{character for the y-label title for plot}

\item{mainTraj}{the main proj plot}

\item{legend}{Logical. Indicates whether the legend prints}

\item{xlimi}{vector bounds to x-axes.}

\item{ylimi}{vector bounds to y-axes.}

\item{panel}{logical if is true, the trajectories are plotted in panel-plot}

\item{colours}{for plot}

\item{...}{additional parameters for plot}
}
\value{
plotted Trajectories Plot of the given DiStatis object.
}
\description{
Plot a \code{TrajPlot} of a DiStatis object
}
\examples{
{
Z2<-DiStatis(NCI60Selec,Scale=TRUE,Center=TRUE)
Colores2<-c(rep(colors()[657],5),rep(colors()[637],6),
rep(colors()[537],7),rep(colors()[552],6),rep(colors()[57],10)
,rep(colors()[300],8),rep(colors()[461],7),rep(colors()[450],2)
,rep(colors()[432],7))
Colores1<-c(rep("Breast",5),rep("CNS",6),rep("Colon",7)
,rep("Leukemia",6),rep("Melanoma",10),rep("Lung",8),rep("Ovarian",7)
,rep("Prostate",2),rep("Renal",7))
TrajPlot(Z2,xlabT = "",ylabT="",colours=Colores2,legend=FALSE)
legend("topleft",unique(Colores1),col=unique(Colores2),bty="n",pch=16,cex=1)

}
}
\author{
M L Zingaretti, J A Demey-Zambrano, J L Vicente Villardon, J R Demey
}

