% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupProj-print.R
\docType{methods}
\name{print.GP}
\alias{print,GroupProj-method}
\alias{print.GP}
\title{\code{\link{print.GP}} a GroupProj object
Generic print.GP method for GroupProj class output visualization.}
\usage{
\S4method{print}{GroupProj}(x)
}
\arguments{
\item{x}{GroupProj class object.}
}
\value{
according to the call
 \item{print}{console generates output text with increasing detail of
 GroupProj object.}
}
\description{
\code{\link{print.GP}} a GroupProj object
Generic print.GP method for GroupProj class output visualization.
}
\examples{
{

 data(NCI60Selec)
 fit<-DiStatis(NCI60Selec,Scale=TRUE,Center=TRUE)
 A<-SelectVar(fit,Crit="p-val(Bonf)",perc=0.95)
 M3<-GroupProj(A,NGroups=4,method="ward",metric="euclidean")

 print(M3)
 summary(M3)

}
}
\author{
M L Zingaretti,  J A Demey-Zambrano, J L Vicente-Villardon, J R Demey
}
\seealso{
\code{\link{GroupProj}}
}

