% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupProj-print.R
\docType{methods}
\name{summary.GP}
\alias{summary,GroupProj-method}
\alias{summary.GP}
\title{\code{\link{summary.GP}} a GroupProj object}
\usage{
\S4method{summary}{GroupProj}(object)
}
\arguments{
\item{object}{GroupProj class object.}
}
\value{
according to the call
 \item{summary}{console output text with increasing detail of
 GroupProj object.}
}
\description{
\code{\link{summary.GP}} a GroupProj object
}
\examples{
{

 data(NCI60Selec)
 fit<-DiStatis(NCI60Selec,Scale=TRUE,Center=TRUE)
 A<-SelectVar(fit,Crit="p-val(Bonf)",perc=0.95)
 M3<-GroupProj(A,NGroups=4,method="ward",metric="euclidean")


 summary(M3)

}
}
\author{
M L Zingaretti, J A Demey-Zambrano, J L Vicente-Villardon, J R Demey
}
\seealso{
\code{\link{GroupProj}}
}

