\name{NCI60Selec_ESet}
\alias{NCI60Selec_ESet}
\docType{data}
\title{Microarray gene expression profiles of the NCI 60 cell lines from 4 different platforms.}
\description{
The 60 human tumour cell lines are derived from patients with leukaemia, melanoma, lung, colon, central nervous system, ovarian, renal, breast and prostate cancers. The cell line panel is widely used in anti-cancer drug screen. In this dataset, a subset of microarray gene expression of the NCI 60 cell lines from four different platforms are combined in a list.
}
\usage{data(NCI60Selec_ESet)}
\format{
    \code{NCI60Selec_ESet} This data is an list of length 4.


}
\value{
NCI60Select_ESet: list of ExpressionSet from four microarrays studies.
}
\source{
Cell Miner http://discover.nci.nih.gov/cellminer/.
}
\references{
Reinhold WC, Sunshine M, Liu H, Varma S, Kohn KW, Morris J, Doroshow J, Pommier Y CellMiner: A Web-Based Suite of Genomic and Pharmacologic Tools to Explore Transcript and Drug Patterns in the NCI-60 Cell Line Set. Cancer Research. 2012 Jul, 15;72(14):3499-511
}
\examples{
data(NCI60Selec_ESet)

}
\keyword{datasets}
