% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectVar-Biplot.R
\docType{methods}
\name{Biplot}
\alias{Biplot}
\alias{Biplot,SelectVar-method}
\title{Plot a \code{\link{Biplot}} of a SelectVar class object}
\usage{
\S4method{Biplot}{SelectVar}(x,xlab=NULL, ylab=NULL, mainP=NULL,
xlimi=NULL, ylimi=NULL, labelObs=TRUE,labelVars= TRUE,
colVar= "black",colObs="black",
 pchPoints=15,Type=c("RMP","CMP","SQRT","HJ"),
 Groups=FALSE,NGroups=2 ,...)
}
\arguments{
\item{x}{DiStatis class object.}

\item{xlab}{character for the x-label title for plot}

\item{ylab}{character for the x-label title for plot}

\item{mainP}{the main Biplot}

\item{xlimi}{(vector) Bounds to x-axis}

\item{ylimi}{(vector) Bounds to y-axis}

\item{labelObs}{Logical. indicates whether the labels of observations are prints. Default is TRUE}

\item{labelVars}{Logical. indicates whether the labels of variables are prints. Default is TRUE}

\item{colVar}{character col for colours of the variables in the plot. Default is black.}

\item{colObs}{character col for colours of the observations in the plot. Default is black.}

\item{pchPoints}{Either an integer specifying a symbol or a single character to be used as the default in plotting points.}

\item{Type}{type of Biplot. Options are CMP RMP SQRT or HJ.}

\item{Groups}{Logical. If is TRUE, the variables are grouped. See \code{\link{GroupProj}}}

\item{NGroups}{Only if the Groups are TRUE. Indicate the number the groups of variables.}

\item{...}{additional parameters for plot}
}
\value{
plotted Biplot/s of the component/s of the given SelectVar object.
}
\description{
Plot a \code{\link{Biplot}} of a SelectVar class object
}
\examples{
{
data(NCI60Selec)
Z1<-DiStatis(NCI60Selec)
M1<-SelectVar(Z1,Crit="R2-Adj")
Colores1<-c(rep("Breast",5),rep("CNS",6),rep("Colon",7),
rep("Leukemia",6),rep("Melanoma",10),rep("Lung",9),
rep("Ovarian",7),rep("Prostate",2),rep("Renal",8))
Colores2<-c(rep(colors()[657],5),rep(colors()[637],6),
rep(colors()[537],7),rep(colors()[552],6),rep(colors()[57],10)
,rep(colors()[300],9),rep(colors()[461],7),rep(colors()[450],2)
,rep(colors()[432],8))
Biplot(M1,labelObs = FALSE,labelVars=FALSE,
colObs=Colores2,Type="SQRT",las=1,cex.axis=0.8,
cex.lab=0.8,xlimi=c(-3,3),ylimi=c(-3,3))
legend("topright",unique(Colores1),col=unique(Colores2),
bty="n",pch=16,cex=0.6)
Biplot(M1,labelObs = FALSE,labelVars=TRUE,colObs=Colores2,
Type="SQRT",las=1,cex.axis=0.8,cex.lab=0.8,xlimi=c(-3,3),
ylimi=c(-3,3),Groups=TRUE,NGroups=6)
legend("topright",unique(Colores1),col=unique(Colores2),
bty="n",pch=16,cex=0.6)
data(winesassesors)
Z3<-DiStatis(winesassesors)
M3<-SelectVar(Z3,Crit="R2-Adj")
Col1<-c(rep("NZ",4),rep("FR",4),rep("CA",4))
Col2<-c(rep(2,4),rep(3,4),rep(4,4))
Biplot(M3,labelObs=FALSE,labelVars=TRUE,colObs=Col2,
Type="SQRT",xlimi=c(-2,2),ylimi=c(-2,2),las=1,cex.axis=0.8,
cex.lab=0.8)
legend("topright",unique(Col1),col=unique(Col2),bty="n",pch=16,cex=0.8)
Biplot(M1,labelObs = FALSE,labelVars=FALSE,colObs=Colores2,
Type="CMP")
legend("topright",unique(Colores1),
col=unique(Colores2),bty="n",pch=16,cex=1)

}
}
\author{
M L Zingaretti,  J A Demey-Zambrano, J L Vicente Villardon, J R Demey
}

