% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectVar-print.R
\docType{methods}
\name{print.SV}
\alias{print,SelectVar-method}
\alias{print.SV}
\title{\code{\link{print.SV}} a SelectVar object
Generic Print/Summary method for SelectVar class output visualization.}
\usage{
\S4method{print}{SelectVar}(x)
}
\arguments{
\item{x}{SelectVar class object.}
}
\value{
according to the call
 \item{print}{the console output the text with increasing detail of
 SelectVar object.}
}
\description{
\code{\link{print.SV}} a SelectVar object
Generic Print/Summary method for SelectVar class output visualization.
}
\examples{
{

 data(NCI60Selec)
 fit<-DiStatis(NCI60Selec,Scale=TRUE,Center=TRUE)
 A<-SelectVar(fit,Crit="p-val(Bonf)",perc=0.95)

 print(A)
 summary(A)

}
}
\author{
M L Zingaretti,  J A Demey-Zambrano, J L Vicente-Villardon, J R Demey
}
\seealso{
\code{\link{SelectVar}}
}

