\encoding{UTF-8}
\name{winesassesors}
\alias{winesassesors}
\docType{data}
\title{Description of a set of wines by a group of expert assessors.}
\description{
Were selected twelve wines made
from Sauvignon Blanc grapes coming from three wine regions (four wines from each region): New Zealand, France, and Canada and then were interviewed 10 expert assessors to evaluate these wines.
}
\usage{data(winesassesors)}
\format{
\code{winesassesors} This data frame contains an list of length 10 with 12 rows each. There are 10 studies, the assessors.
}
\value{
List with 10 items  (all data.frame)  with 12 rows each (common elements in Statis Methods) and differents variables types.
}
\source{
Abdi, H., Williams, L.J., Valentin, D., & Bennani-Dosse, M. (2012). STATIS and DISTATIS: optimum multitable principal component analysis and three way metric multidimensional scaling. WIREs Comput Stat, 4, 124-167.
}
\references{
Abdi, H., Williams, L.J., Valentin, D., & Bennani-Dosse, M. (2012). STATIS and DISTATIS: optimum multitable principal component analysis and three way metric multidimensional scaling. WIREs Comput Stat, 4, 124-167.
}
\examples{
data(winesassesors)

}
\keyword{winesassesors}
